fix: block 'session-changed' for a duration
We were depending on each element to send a 'session-changed' event to
consider that the load was complete, however this is an incorrect use of
the 'session-changed' event. It is designed to only fire when the user
has made a change, any element that is firing when the session is set is
using it incorrectly. Because we can't depend on the elements behaving
better in this regard, just block any event from propagating for a short
duration when the model/sessions are being set.

TODO: is 'loadComplete' in use? If so it should probably be looking at
the model set event for each el instead?