fix: authoring view not loading PIE-155
This temoporary fix disables the use of
pie_registry when using `loadPlayerJson` (which will have a bundle.url property in model) this
reverts to how things were before. But this
should be changed to use an enhanced registry
to help optimize and not double-load PIEs.

The bug was that the registry thought it had
loaded the PIE but the bundle was player.js
which does not have Config constructor