feat: normalize elements
The pie config spec lets a user map packages
with element names, models and markup.
For the purposes of browser rendering we
want to control the element tag names used to
allow control over custom element management / loading.
This fixes:ch6835 which was caused due to
different tag names for same element.